# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  handlers.SearchRequestBody:
    properties:
      collection:
        type: string
      filter:
        additionalProperties: true
        type: object
      min_score:
        type: number
      query:
        type: string
      top_k:
        type: integer
      use_cache:
        type: boolean
    type: object
...
  services.SearchResponse:
    properties:
      collection:
        type: string
      embedding_time_ms:
        type: number
      from_cache:
        type: boolean
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/services.SearchResult'
        type: array
      search_time_ms:
        type: number
      total_results:
        type: integer
    type: object
  services.SearchResult:
    properties:
      chunk_id:
        type: string
      distance:
        type: number
      document_id:
        type: string
      document_name:
        type: string
      metadata:
        additionalProperties: true
        type: object
      score:
        type: number
      text:
        type: string
    type: object
  services.UploadDocumentResponse:
    properties:
      chunk_count:
        type: integer
      collection:
        type: string
      document_id:
        type: string
      filename:
        type: string
      job_id:
        type: string
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      processing_time_ms:
        type: number
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for document processing, analysis, and risk assessment
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Risk Analyzer API
  version: "1.0"
paths:
...
  /api/v1/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload and process a document for vector storage
      parameters:
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      - description: Collection name
        in: formData
        name: collection
        required: true
        type: string
      - default: semantic
        description: Chunking strategy
        in: formData
        name: chunking_strategy
        type: string
      - default: 512
        description: Chunk size
        in: formData
        name: chunk_size
        type: integer
      - default: 50
        description: Chunk overlap
        in: formData
        name: chunk_overlap
        type: integer
      - default: false
        description: Extract metadata
        in: formData
        name: extract_metadata
        type: boolean
      - default: 3
        description: Number of questions
        in: formData
        name: num_questions
        type: integer
      - default: 0
        description: Max pages to process
        in: formData
        name: max_pages
        type: integer
      - default: false
        description: Process asynchronously
        in: formData
        name: async
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.UploadDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upload a document
      tags:
      - documents
  /api/v1/search:
    get:
      description: Perform a simple search using query parameters
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Collection name
        in: query
        name: collection
        required: true
        type: string
      - default: 10
        description: Number of results
        in: query
        name: top_k
        type: integer
      - default: true
        description: Use cache
        in: query
        name: use_cache
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Simple search
      tags:
      - search
    post:
      consumes:
      - application/json
      description: Perform vector similarity search across documents
      parameters:
      - description: Search request
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/handlers.SearchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Search documents
      tags:
      - search
...
  /documents/chunks:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Legacy endpoint. Get chunks from the vector store
        with pagination'
      parameters:
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      - description: Filter by document ID
        in: query
        name: document_id
        type: string
      - default: 100
        description: Maximum number of chunks
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of chunks to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Get document chunks'
      tags:
      - deprecated
  /documents/collection-stats:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/collections/{name}/stats instead. Get
        statistics about the vector store collection'
      parameters:
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Get collection statistics'
      tags:
      - deprecated
  /documents/list:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/documents instead. Get a list of all documents
        stored in the vector database'
      parameters:
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] List all documents'
      tags:
      - deprecated
  /documents/process-example:
    post:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Legacy endpoint. Process the bundled example PDF through
        the full pipeline'
      parameters:
      - default: sentence
        description: Chunking strategy
        in: query
        name: chunking_strategy
        type: string
      - default: 512
        description: Chunk size
        in: query
        name: chunk_size
        type: integer
      - default: 50
        description: Chunk overlap
        in: query
        name: chunk_overlap
        type: integer
      - default: true
        description: Store in vector DB
        in: query
        name: store_in_vector_db
        type: boolean
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Process example PDF'
      tags:
      - deprecated
  /documents/upload:
    post:
      consumes:
      - multipart/form-data
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/documents/upload instead. Upload a PDF
        or document file and process it'
      parameters:
      - description: Document file to upload
        in: formData
        name: file
        required: true
        type: file
      - default: sentence
        description: Chunking strategy
        in: query
        name: chunking_strategy
        type: string
      - default: 512
        description: Chunk size
        in: query
        name: chunk_size
        type: integer
      - default: 50
        description: Chunk overlap
        in: query
        name: chunk_overlap
        type: integer
      - default: false
        description: Store in vector DB
        in: query
        name: store_in_vector_db
        type: boolean
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Upload and process document'
      tags:
      - deprecated
...
  /search:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/search instead. Search for similar chunks
        in the vector store'
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      - default: 5
        description: Number of results
        in: query
        name: top_k
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Search documents'
      tags:
      - deprecated
...
  /search/collections/{collection_name}:
    delete:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/collections/{name} DELETE instead. Delete
        and reset a collection'
      parameters:
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Reset collection'
      tags:
      - deprecated
  /search/collections/{collection_name}/stats:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/collections/{name}/stats instead. Get
        statistics for a specific collection'
      parameters:
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Get collection statistics'
      tags:
      - deprecated
  /search/query:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/search instead. Perform a quick search
        across documents'
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 5
        description: Number of results
        in: query
        name: top_k
        type: integer
      - description: Collection name
        in: query
        name: collection
        type: string
      - description: Filter by document ID
        in: query
        name: document_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Quick search documents'
      tags:
      - deprecated
swagger: "2.0"